fixed getData functioons in MyDatabase. SitesActivity needs fix
I added getSitesData and getMapsData function with the correct query in
MyDatabase, altho SitesActivity still have some legacy code that needs
to be cleaned.
The method I used is adding file io (i know the dumb way but effective).
I grabbed some code from internet and it perfectly work. Our database
doesnt need to be created in the code, so I shudnt do it. Instead I need
to check if the database already exist and update it with the db file in
assets.